+++ src/main/scala/org/scalasbt/jacoco/report/Reporting.scala
@@ -10,36 +10,12 @@
  * http://www.eclipse.org/legal/epl-v10.html
  */
 
-package org.scalasbt.jacoco
+package org.scalasbt.jacoco.report
 
-import java.io.FileOutputStream
-
-import org.jacoco.core.data.ExecutionDataWriter
-import org.jacoco.core.runtime.RuntimeData
-import org.scalasbt.jacoco.report.{JacocoReportSettings, JacocoSourceSettings}
 import sbt.Keys._
 import sbt._
 
 import scala.language.implicitConversions
-
-private[jacoco] object SavingData {
-  def saveDataAction(jacocoFile: File, forked: Boolean, streams: TaskStreams)(
-      implicit runtimeData: RuntimeData): Unit = {
-
-    if (!forked) {
-      IO createDirectory jacocoFile.getParentFile
-      val executionDataStream = new FileOutputStream(jacocoFile)
-      try {
-        streams.log debug ("writing execution data to " + jacocoFile)
-        val executionDataWriter = new ExecutionDataWriter(executionDataStream)
-        runtimeData collect (executionDataWriter, executionDataWriter, true)
-        executionDataStream.flush()
-      } finally {
-        executionDataStream.close()
-      }
-    }
-  }
-}
 
 private[jacoco] object Reporting {
   def reportAction(
